move all dialogs to dialogs.js, add utility for viewing them

+++ data/ui.js
@@ -97,7 +97,7 @@ function confirmStartSharing(askForConfirmation,elmClicked){
 }
 
 function confirmStopSharing(elmClicked){
-     stopSharing(function(confirmed){
+     stopSharingDialog(function(confirmed){
         if ( confirmed ){
             toggleBtnOffEffect( document.querySelector(".share-btn") );
         }else{
@@ -150,24 +150,20 @@ document.querySelector(".download").addEventListener('click', function(evt) {
     // window.open('data:application/json,' + exportFormat(allConnections));
 });
 
+
+
 document.querySelector('.reset-data').addEventListener('click', function(){
-    dialog( {   "name": dialogNames.resetData,
-                "title": "Reset Data",
-                "message":  "<p>Pressing OK will delete all Lightbeam information including connection history, user preferences, unique token, block sites list etc.</p>" + 
-                            "<p>Your browser will be returned to the state of a fresh install of Lightbeam.</p>",
-                "imageUrl": "image/collusion_popup_warningreset.png"
-            },function(confirmed){
-                if ( confirmed ){
-                    // currentVisualization.emit('remove');
-                    allConnections = [];
-                    addon.emit('reset');
-                    aggregate.emit('reset');
-                    userSettings = {};
-                    localStorage.clear();
-                    location.reload(); // reload page
-                }
-            }
-    );
+    confirmResetDataDialog(function(confirmed){
+        if ( confirmed ){
+            // currentVisualization.emit('remove');
+            allConnections = [];
+            addon.emit('reset');
+            aggregate.emit('reset');
+            userSettings = {};
+            localStorage.clear();
+            location.reload(); // reload page
+        }
+    });
 });
 
 // function handleDisclosureToggle(elem){
@@ -521,6 +517,8 @@ function resetHighlightedRow(){
 *   Special dialog handler for promptToShareDataDialog
 *   FIXME: temporary solution for now.  need to clean up the code a bit.
 */
+
+
 const promptToShareDialogShowLimit = 3;
 function showPromptToShareDialog(){
     var showTimes, today, shownToday, belowLimit;
@@ -532,23 +530,13 @@ function showPromptToShareDialog(){
     if ( localStorage.numLaunch > 1 && !doNotShowDialog(dialogNames.promptToShare) && !shownToday && belowLimit){
         showTimes.push(today);
         localStorage.promptToShareDialogShowTimes = JSON.stringify(showTimes);
-        dialog( {
-                "name": dialogNames.promptToShare,
-                "dnsPrompt": true,
-                "title": "Help the Ecosystem by Sharing",
-                "message":  "<p>As a user of Lightbeam Beta, you can help contribute to build our data ecosystem.</p>" + 
-                            "<p>By sharing your data you can help us and others to understand third-party relationships on the web and promote further research in the field of online tracking and privacy.</p>  "+
-                            "<p>Do you want to upload your de-identified data to the public database now?</p>",
-                "imageUrl": "image/collusion_popup_startsharing.png"
-            },
-            function(confirmed){
-                if( confirmed ){
-                    var sharingToggle = document.querySelector(".toggle-btn.share-btn input");
-                    confirmStartSharing(false,sharingToggle);
-                    disablePromptToShareDataDialog();
-                }
+        showPromptToShareDialog(function(confirmed){
+            if( confirmed ){
+                var sharingToggle = document.querySelector(".toggle-btn.share-btn input");
+                confirmStartSharing(false,sharingToggle);
+                disablePromptToShareDataDialog();
             }
-        );
+        });
     }
 }
